+++ Changelog.md
@@ -2,6 +2,7 @@
  - Redirect to main-menu after the password has changed.
  - Fix registration without gender.
  - Fix spacing of some elements.
+ - Add filters (mealtime, difficulty) to recipe-info page.
 
  
 ## [1.1.0] - 2020-02-06